pyproject.toml
[tool.poetry]
name = "plotit"
version = "0.1.0"
description = "Plot SVG's :)"
authors = ["Maarten Hoogendoorn <maarten.hoogendoorn@navara.nl>"]
license = "GPL3.0"

[tool.poetry.dependencies]
python = "^3.8"
svg-to-gcode = "^1.5.4"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
